requirements.txt
aiogram>=3.0.0
SQLAlchemy==1.4.46
python-dotenv==0.21.0